+++ src/ZenExtensions.Slack.Models/CompositionObjects/TextObject.cs
@@ -12,13 +12,14 @@ namespace ZenExtensions.Slack.Models.CompositionObjects
     ///     The text for the block. This field accepts any of the standard text formatting markup when type is mrkdwn. Maximum length for the text in this field is 3000 characters.
     /// </param>
     public abstract record class TextObject(
-        [property: JsonPropertyName("text")]
+        [property: JsonPropertyName("text"), JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
         string Text
     ) : CompositionObject
     {
         /// <summary>
         ///    The formatting to use for this text object. Can be one of plain_textor mrkdwn.
         /// </summary>
+        [JsonIgnore(Condition = JsonIgnoreCondition.Always)]
         public abstract string Type { get; protected internal set; }
     }
 }