+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [1.1.0] - 2023-01-26
+
 ### Added 
 
 - Added slack [composition objects](https://api.slack.com/reference/block-kit/composition-objects)
+- Added json ignore condition to when writing null
 
 ### Changed
 
@@ -24,5 +28,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Addex slack slash request models
 - Added http extensions for slack message
 
-[Unreleased]: https://github.com/ZenExtensions/slack-models/compare/1.0.0...HEAD
+[Unreleased]: https://github.com/ZenExtensions/slack-models/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/ZenExtensions/slack-models/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/ZenExtensions/slack-models/releases/tag/1.0.0